View accounts summary

+++ Makefile
@@ -1,18 +1,14 @@
 emacs ?= emacs
 
-.PHONY: all
-all: build
+.PHONY: clean
+clean:
+	if [ -d build ]; then rm -r build; fi
+
+.PHONY: build
+build: clean
+	@mkdir -p build
+	@cat *.el > build/marco-polo.el
 
 .PHONY: run
 run: build
 	$(emacs) -q -l build/marco-polo.el
-	make clean
-
-.PHONY: build
-build:
-	@mkdir -p build
-	@cat *.el > build/marco-polo.el
-
-.PHONY: clean
-clean:
-	rm -r build
